Created common component for Bonus and resolved issues

+++ src/components/NoteItem.js
@@ -148,15 +148,17 @@ class NoteItem extends Component {
   };
 
   toggleDeleteConfirmationModal = (index = null, item = null) => {
-    this.setState((prevState) => ({
-      showDeleteConfirmationModal: !prevState.showDeleteConfirmationModal,
-      deleteIndex: index,
-      deleteItem: item,
-    }));
+        this.setState((prevState) => ({
+            showDeleteConfirmationModal: !prevState.showDeleteConfirmationModal,
+            deleteIndex: index,
+            deleteItem: item,
+        }));
   };
 
   onConfirmDelete = () => {
-    this.deleteComment(this.state.deleteItem.id);
+    if (this.state.deleteItem && this.state.deleteItem.id){
+        this.deleteComment(this.state.deleteItem.id);
+    }
   };
 
   getCommentList = (note_id, offset) => {
